### 2019-10-12 #1804 @labarba
@trallard 👋  — Could you handle this submission as the editor?
**mwarp1d: Manual one-dimensional data warping in Python and PyQt**
https://github.com/0todd0000/mwarp1d
(The author suggested @Kevin-Mattheus-Moerman, but he has a high load of submissions right now.)
