### 2019-10-13 cd744e2
Update paper.md
