### 2019-10-13 #1804 @0todd0000
Comment re: article proof:

Figure problem:  the figure was not visible in the proof (using HTML tags) so I've updated paper.md in the repository using Markdown image syntax according to JOSS instructions.

Apologies for my mistake!

I have also corrected two other minor errors in the proof.

Please re-compile the article proof at your convenience.
