### 2019-10-13 #1804 @0todd0000
The proof looks OK now.
