### 2019-10-23 #1804 @trallard
Hi @0todd0000  the first thing we have to do to move this review forward is finding a pair of suitable reviewers. Have you got someone in mind already?

You can start by finding some reviewers looking at our list of registered reviewers here <https://bit.ly/joss-reviewers>
