### 2019-11-05 #1804 @trallard
Thanks for the ping @0todd0000

---
Dear potential reviewers,

I am looking for a couple of folks willing and able to review the following JOSS submission:

**Title**: mwarp1d: Manual one-dimensional data warping in Python and PyQt
**description**: Manual one-dimensional data warping and nonlinear registration in Python and PyQt
**repository**: https://github.com/0todd0000/mwarp1d
**documentation**: http://www.spm1d.org/mwarp1d/index.html
**JOSS paper**: <https://github.com/openjournals/joss-papers/blob/joss.01804/joss.01804/10.21105.joss.01804.pdf>

@melund, Morten are you willing and able to take upon this review?
@CELAvecchia, Carolina are you willing and able to take upon this review?
@alcantarar, Ryan are you willing and able to take upon this review?
@abhishektha, Abhishekta are you willing and able to take upon this review?
@ ivoflipse Ivo, are you willing and able to take upon this review?
@ demotu Marcos, are you willing and able to take upon this review?

We are looking for at least two reviewers. The reviewing guidelines for JOSS can be found at <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>
