### 2019-11-07 #1870 @whedon
[REVIEW]: mwarp1d: Manual one-dimensional data warping in Python and PyQt
**Submitting author:** @0todd0000 (<a href="http://orcid.org/0000-0002-8292-7189">Todd Pataky</a>)
**Repository:** <a href="https://github.com/0todd0000/mwarp1d" target ="_blank">https://github.com/0todd0000/mwarp1d</a>
**Version:** 0.2
