### 2019-11-25 0todd0000/mwarp1d#1 @melund
[JOSS Review] Extending the readme file
It think all the formal JOSS requirements for documenting  contributing guidelines, license etc. are all fulfilled in the documentation page.

However, I would still suggest you make some of these things a little easier to find by extendíng the readme page in the main repository. Many users will probably find the software first on GitHub, and would be nice if the readme was more informative.

See these examples for awesome readme files:
https://github.com/matiassingers/awesome-readme
