### 2019-11-25 0todd0000/mwarp1d#2 @melund
[JOSS Review] What about automated tests?
The minimal  JOSS requirements testing reads:

> Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)

I guess you could basically argued that some of the [steps in the documentation](http://www.spm1d.org/mwarp1d/examples.html) full fills this requirement. But there is not really any objective criteria to say if the warping was correct. In other words how can I know the library does the same as the last release.

I think you could benefit from automated testing. Tests are also a nice if you ever get contributions from others as it helps verify their code. Using something like "Github actions" is very easy to setup,
