### 2019-11-25 0todd0000/mwarp1d#3 @melund
[JOSS Review] Include source code for documentation. 
Where is the source code for the documentation? I think it would be good practice if the sphinx source code for the documentation lived in the repository.

Contributing to documentation is also big part of "open source", and it would allow you automatically build the documentation with a CI service like github actions.
