### 2019-11-27 #1870 @0todd0000
@melund Oops!  I didn't see the issues, thank you for pointing them out. Will get on them now.
