### 2019-11-27 0todd0000/mwarp1d#2 @0todd0000
Since the package's main purpose is to support manual (and therefore subjective) registration, the only way to thoroughly automate quantitative testing is to simulate mouse and keyboard actions.  This is certainly possible using Qt's event simulators, but this type of automated testing takes much longer to develop than normal script-based testing. Given the JOSS requirement statement (i.e., that step-by-step documentation is sufficient for publication), we decided to choose the documentation route instead.

Please note:
The documentation link you sent above (["steps in the documentation"](http://www.spm1d.org/mwarp1d/examples.html)) contains only a list of scripting examples.

The step-by-step instructions appear in the [User manual](http://www.spm1d.org/mwarp1d/usermanual.html) section, and for those who don't like to read, in the [Screencasts](http://www.spm1d.org/mwarp1d/quickstart.html) section.

Do you think Qt-based mouse+keyboard automation is necessary for publication?
