### 2019-11-27 #1870 @melund
No problem. I am in general positive about your tool. I have a little difficult seeing the use case for the the full manual warping. Maybe I am missing something. Is it possible to warp multiple segments along the curve or is it only one setting for the whole curve?  In what cases would you use it yourself?

I do, however, think that a nice GUI tool for making landmark based warping could be really useful. Are there any ongoing projects in your group that use or plan to use (and cite) this tool? I think it would be good to mention them in the paper.

So, in conclusion, if you address the issue I have about missing documentation source and lacking tests, then I think I can recommend this  publication.
