+++ paper/paper.bib
@@ -1,110 +1,0 @@
-
-@article{Buckner:2004,
-    author = {Buckner, Randy L and Head, Denise and Parker, Jamie and Fotenos, Anthony F and Marcus, Daniel and Morris, John C and Snyder, Abraham Z},
-    title = {{A unified approach for morphometric and functional data analysis in young, old, and demented adults using automated atlas-based head size normalization: reliability and validation against manual measurement of total intracranial volume}},
-    journal = {NeuroImage},
-    year = {2004},
-    volume = {23},
-    number = {2},
-    pages = {724--738},
-    doi = {10.1016/j.neuroimage.2004.06.018},
-}
-
-
-
-
-@article{Iosifescu:1997,
-    author = {Iosifescu, Dan V and Shenton, Martha E and Warfield, Simon K and Kikinis, Ron and Dengler, Joachim and Jolesz, Ferenc A and McCarley, Robert W},
-    title = {{An Automated Registration Algorithm for Measuring MRI Subcortical Brain Structures}},
-    journal = {NeuroImage},
-    year = {1997},
-    volume = {6},
-    number = {1},
-    pages = {13--25},
-    doi = {10.1006/nimg.1997.0274},
-}
-
-
-
-@article{Marron:2015,
-    author = {Marron, J S and Ramsay, James O and Sangalli, Laura M and Srivastava, Anuj},
-    title = {{Functional data analysis of amplitude and phase variation}},
-    journal = {Statistical Science},
-    year = {2015},
-    volume = {30},
-    number = {4},
-    pages = {468--484},
-    month = nov,
-    doi = {10.1214/15-STS524},
-}
-
-
-
-@article{Qin:2009,
-    author = {Qin, A K and Huang, V L and Suganthan, P N},
-    title = {{Differential Evolution Algorithm With Strategy Adaptation for Global Numerical Optimization}},
-    journal = {IEEE Transactions on Evolutionary Computation},
-    year = {2009},
-    volume = {13},
-    number = {2},
-    pages = {398--417},
-    month = mar,
-    doi = {10.1109/TEVC.2008.927706},
-}
-
-
-
-@article{Ramsay:1998,
-    author = {Ramsay, J O and Li, X},
-    title = {{Curve registration}},
-    journal = {Journal of the Royal Statistical Society Series B},
-    year = {1998},
-    doi = {10.1111/1467-9868.00129},
-}
-
-
-@book{Ramsay:2005,
-    author = {J. O. Ramsay and B. W. Silverman},
-    title = {{Functional Data Analysis}},
-    publisher = {Springer-Verlag},
-    address={New York},
-    year = {2005},
-    url = "http://www.psych.mcgill.ca/misc/fda/downloads/",
-}
-
-
-@article{Sadeghi:2000,
-    author = {Sadeghi, H and Allard, P and Shafie, K and Mathieu, PA and Sadeghi, S and Prince, F and Ramsay, J},
-    title = {{Reduction of gait data variability using curve registration}},
-    journal = {Gait and Posture},
-    year = {2000},
-    volume = {12},
-    number = {3},
-    pages = {257--264},
-    doi = {10.1016/S0966-6362(00)00085-0},
-}
-
-
-
-@article{Wrobel:2018,
-    author = {Wrobel, J},
-    title = {{register: Registration for Exponential Family Functional Data}},
-    journal = {The Journal of Open Source Software},
-    year = {2018},
-    volume = {3},
-    number = {22},
-    pages = {557},
-    url = "https://github.com/julia-wrobel/registr",
-    doi = {10.1111/biom.12963},
-}
-
-
-@article{Zeng:2019,
-    author = {P. Zeng and S. J. Qing and W. S. Kim},
[…]
