### 2019-11-27 0todd0000/mwarp1d#3 @0todd0000
Good idea, sorry about that!
All Sphinx source code has been uploaded to the [gh-pages](https://github.com/0todd0000/mwarp1d/tree/gh-pages) branch.
