### 2019-11-27 0todd0000/mwarp1d#3 @melund
Nice. I didn't know gh-pages could also build sphinx documentation.

Just keep in mind that having docs in a separate branch can make it a little more difficult to sync source and documentation. It is not a blocker for the review but you could move the docs source  into a docs/ subfolder in the main repository. You can still exclude the docs source when you build pypi and conda packages.

I think gh-pages can also handle that the [docs is located in a sub folder](https://www.docslikecode.com/articles/github-pages-python-sphinx/)  in the main branch..
