### 2019-11-27 0todd0000/mwarp1d#1 @0todd0000
Agreed, thank you for the suggestion. I've replicated most of the documentation's front page in the repository's README file, including links to the main documentation sections.
