### 2019-11-27 #1870 @0todd0000
Thanks very much for the feedback!

(I've just responded to the three repository issues, including the Sphinx source code and automation. I had notifications turned off for some reason, so didn't see the issues before today, sorry about that!)

> I have a little difficult seeing the use case for the the full manual warping.

For landmark warping, the warp resolution (and indirectly: warp quality) is directly related to the number of landmarks. The manual mode permits high-resolution warping using mouse drags rather than mouse clicks.

> Is it possible to warp multiple segments along the curve or is it only one setting for the whole curve?

Yes, independent manual warps can be sequentially applied as indicated in the [manual warp screencast](http://www.spm1d.org/mwarp1d/quickstart.html#manual-warping-screencast) (at  around 0:40 ). These manual warps can affect an arbitrarily large (or small) regions of the 1D domain. In order to get finer control over local regions, the data must  be pre-interpolated to a finer resolution (e.g. 1000 points instead of 100).

> In what cases would you use it yourself?

We're not yet totally sure, but we're trying to answer this question objectively through experimentation. We're currently running a study to check performance differences between different methods (landmark, manual with low-resolution, manual with high-resolution), to try to identify the dataset characteristics that are best for each registration mode.

 > Are there any ongoing projects in your group that use or plan to use (and cite) this tool? I think it would be good to mention them in the paper.

Yes, the project above, but it's not yet close to a paper. We will certainly add it as a reference in the documentation if and when the study is published.
