### 2019-11-28 #1870 @demotu
Hi Todd
The application is very interesting and potentially useful.
It's unfortunate that you don't have mwarp1d for Linux in your conda channel yet. So, you could mention that in your installation page and just tell the user what he/she already knows: install with python setup.py install.
The user manual and the example in overview.ipynb are helpful and clear.
The application is very complete and give lots of options for performing data warping.
The references don't appear in the paper.md but I can see them in the bib file, I guess it's ok.
Otherwise, as a biomechanicist working with these kind of problems, I didn't have any difficult in using your application.
Congratulations.
