### 2019-11-28 #1870 @melund
>Yes, independent manual warps can be sequentially applied as indicated in the manual warp screencast (at around 0:40 ).

Thanks @0todd0000. I missed that when I played with the program.
