### 2019-11-28 0todd0000/mwarp1d#2 @melund
No. Please don't do any automated GUI testing. That would be overkill for an application like this. It is the underlying logic that could benefit from unit testing.

You have plenty of scripting examples in `http://www.spm1d.org/mwarp1d/examples.html`, which could easily be tested. Functions like `mwarp1d.warp_manual`

I agree that JOSS does not require unit testing, and that your documentation is extensive enough comply with the minimum JOSS requirements. So I guess this is not a blocker for the review.

I do understand that you don't see the benefit of automated testing if you haven't used it in other projects. But for open source projects automated CI testing is often quite essential. They can help ensure that your stuff still works when dependencies change over time. How can you be sure that the program will still load with Python 4 when it has been lying around untouched for three years.

For example, your `Setup.py` and Conda recipe does not mention all the dependencies that you use. This is guaranteed to break at some point when the default QT in Anaconda changes. I think you would have caught this error if you did automated testing.

Also,  you should consider creating a conda-forge recipe. It will ensure that the package will continue to work in the future and also give you packages for linux.
