### 2019-11-28 0todd0000/mwarp1d#3 @melund
I think I can close this. My remaining comments are not blockers.
