### 2019-11-28 #1870 @melund
@0todd0000 has handled or answered all of my important comments:

https://github.com/0todd0000/mwarp1d/issues/1
https://github.com/0todd0000/mwarp1d/issues/2
https://github.com/0todd0000/mwarp1d/issues/3

@trallard I can also recommend this for publication.
