### 2019-11-29 0todd0000/mwarp1d#2 @0todd0000
Thank you for that! I do use unit testing in other projects, I just usually keep the unit testing code separate from the main repository. But I agree it could indeed be advantageous to include the unit testing code, will definitely add it in the future. Thank you!

I've not used conda-forge before, but will definitely try it out, thanks for that as well!
