### 2019-12-01 #1870 @0todd0000
@trallard I think both reviewers have recommended publication. Is there anything else I or the reviewers need to do for the review process?
