### 2019-12-05 #1870 @trallard
Hey @0todd0000  I will start with my final editor duties. In the meantime, can you please complete the following actions?

- Final review of the paper for typos, author affiliations and missing DOIs
- Create a final release of the package and archive (i.e. Zenodo, Figshare)
- Indicate the final version of the package here
- Share the newly generated DOI here
