+++ paper/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 
 Warping 1D data has been shown to be necessary for minimizing variability in a variety of applications [@Ramsay:1998; @Sadeghi:2000; @Marron:2015], yet only a few open-source algorithmic techniques exist [@Ramsay:2005; @Wrobel:2018; @Zeng:2019], and no manual techniques exist for 1D registration. It has separately been shown that, for 2D data, manual registration can perform as well as, and in some cases better than algorithmic registration [@Iosifescu:1997; @Buckner:2004]. Manual registration may also be necessary when algorithmic registration fails or requires manual interaction [@Qin:2009].
 
-``mwarp1d`` is a tool for manually warping 1D data via Python scripting and also via a PyQt graphical user interface (GUI). ``mwarp1d`` implements two warping procedures: `landmark` and `manual`. Landmark registration was introduced to the literature more than 20 years ago, involving homologous point digitizing and subsequent piecewise interpolation. While simple, landmark registration was implemented as a comparative baseline. The `manual` registration scheme is novel, involving domain warps through manually adjustable, globally-constrained asymmetric Gaussian warping kernels.
+``mwarp1d`` is a tool for manually warping 1D data via Python scripting and also via a PyQt graphical user interface (GUI). ``mwarp1d`` implements two warping procedures: `landmark` and `manual`. Landmark registration was introduced to the literature more than 20 years ago, involving homologous point digitizing and subsequent piecewise interpolation. While simple, landmark registration was implemented in ``mwarp1d`` as a comparative baseline. The `manual` registration scheme is novel, involving domain warps through manually adjustable, globally-constrained asymmetric Gaussian warping kernels.
 
 
 # Statement of need
@@ -42,7 +42,7 @@ Warping 1D data has been shown to be necessary for minimizing variability in a v
 
 **Problem solved**:  `mwarp1d` allows users to manually define arbitrarily complex warps, thereby solving the problem of potentially poor algorithmic registration performance. As far as we are aware, no other manual, GUI-based software exists for manual 1D data warping.
 
-**Target audience**: The target audience is anyone working with or analyzing registerable 1D data, especially researchers in the human and animal movement sciences.
+**Target audience**: The target audience is anyone working with or analyzing registerable 1D data, especially researchers in the human and animal movement sciences. Full documentation is available in the [project repository](https://github.com/0todd0000/mwarp1d).
 
 # Figures
 
