### 2019-12-10 #1870 @0todd0000
@trallard

All final tasks complete. Here are the details:

* Final review of the paper for typos, author affiliations and missing DOIs

    * Two minor additions made to the paper for clarity. No typos found.
    * Author affiliations are correct.
    * No missing DOIs;  all DOIs re-checked.

* Create a final release of the package and archive (i.e. Zenodo, Figshare)

    * Done (Zenodo).

* Indicate the final version of the package here

     * Version:  0.2

* Share the newly generated DOI here:

    * [![DOI](https://zenodo.org/badge/190681079.svg)](https://zenodo.org/badge/latestdoi/190681079)
