### 2019-12-10 #1870 @0todd0000
@trallard
The updated proof looks good.
