### 2019-12-11 #1870 @0todd0000
@trallard
Oops!  Sorry about that, the author list is now updated at Zenodo.
