### 2019-12-14 #1870 @trallard
I believe that this is ready to pass over to the last stage of the editorial process.

@melund @demotu, thanks a lot for your time and contributions to JOSS through this review 🙌🏼

@openjournals/joss-eics I have concluded my editorial tasks and I would recommend this submission for publication. Thus I am sending it your way 👾🚀
