### 2019-12-15 #1870 @kyleniemeyer
Hi @0todd0000, I'm just doing the final checks on the submission before publishing. Almost everything looks good, but I noticed one of your references was missing some info—could you add the volume, issue, and page number to the Ramsay and Li (1998) reference? I found: Volume 60, Issue 2, pages 351-363. Thanks!
