+++ paper/paper.bib
@@ -59,6 +59,9 @@
     title = {{Curve registration}},
     journal = {Journal of the Royal Statistical Society Series B},
     year = {1998},
+    volume = {60},
+    number = {2},
+    pages = {351--363},
     doi = {10.1111/1467-9868.00129},
 }
 
