### 2019-12-16 #1870 @0todd0000
Hi @kyleniemeyer, sorry about that!  I've updated the reference and have regenerated the PDF.
