### 2019-12-16 #1870 @kyleniemeyer
Thanks to @melund and @demotu for reviewing, and @trallard for editing.
